Feat/conv self test
* fix: fix agent builder ur


* fix: fix ut


* fix: agent chat  test & bugfix


* fix: agent chat  test & bugfix


* fix: fix field


* fix: fix


* fix: end-to-end test


* fix: end-to-end test


See merge request: !244

+++ backend/domain/conversation/conversation/conversation_impl.go
@@ -81,7 +81,7 @@ func (c *conversationImpl) NewConversationCtx(ctx context.Context, req *entity.N
 
 func (c *conversationImpl) GetCurrentConversation(ctx context.Context, req *entity.GetCurrentRequest) (*entity.Conversation, error) {
 	// get conversation
-	conversation, err := c.ConversationDAO.Get(ctx, req.UserID, req.AgentID, req.Scene)
+	conversation, err := c.ConversationDAO.Get(ctx, req.UserID, req.AgentID, int32(req.Scene))
 
 	if err != nil {
 		return nil, err
